+++ Project/demo/src/main/java/demo/demo/service/StudentService.java
@@ -7,7 +7,7 @@ import demo.demo.model.Student;
 public interface StudentService {
 	Student saveStudent(Student student);
 	List<Student> getAllStudent();
-	Student getStudentById(long id);
-	Student updateStudent(Student student, long id);
-	void deleteStudent(long id);
+	Student getStudentById(Integer id);
+	Student updateStudent(Student student, Integer id);
+	void deleteStudent(Integer id);
 }
